CHore: rename utils file
