+++ src/components/DataRoom/Hero/utils/AsciiEffects.ts
